requirements.txt
Jinja2==2.9.5
MarkupSafe==0.23
PyYAML==3.12
appdirs==1.4.0
beautifulsoup4==4.5.3
bs4==0.0.1
html5lib==0.999999999
packaging==16.8
pyparsing==2.1.10
six==1.10.0
webencodings==0.5
